+++ PruningMarioResearch/src/graphBuilder/GraphBuilder.java
@@ -66,7 +66,7 @@ import dk.itu.mario.level.Level;
 public class GraphBuilder
 {
 
-	private int counterIDs;
+	private long counterIDs;
 	private int counterBranches;
 	private ArrayList<BlockNode> Beststates;
 	private ArrayList<BlockNode> Worststates;
@@ -382,7 +382,7 @@ public class GraphBuilder
     public boolean  validationSymmetryFuture(   int countElements, int countElementsFinal,ArrayList statesCopy, ArrayList finalList, ElementsToPlace objElemP,  Random random, double partialSymmetry,int floorTileHeight,int height)
     { 
     	
-    	int counterIDsCopy=counterIDs;
+    	long counterIDsCopy=counterIDs;
     	countElements=countElements-1;
     	//Collections.sort(bestXs,Collections.reverseOrder());
     	//Collections.sort(bestYs,Collections.reverseOrder());
@@ -4187,7 +4187,7 @@ public class GraphBuilder
 		return summatory;
 	}
 	
-	public int getCounterIDs()
+	public long getCounterIDs()
     {
     	return counterIDs;
     }
